Fixed: Store current record id in relationship
Since the app serializers are only handed the embedded JSON and not the full record for the relationships, they do not have access to the record type and id of the 'parent' record. We could pass this along in the serializer context, but it's much easier and less prone to failure to just add it to the index document.