New: Add date parsing library
This commit adds a new date parsing library that tries its very best to parse a date out of the wide range of date statements we have in the records. It doesn't always succeed, but it's pretty good.

Installs edtf from the rism-digital fork since it contains a significant performance enhancement.