Fixed: issues with pre-1000 dates
If a date was listed as being pre-1000 (e.g., `701-762`) we would previously consider this to be a post-millenium date with a missing year, and it was translated to the middle of the decade, e.g., "197?" -> "1975".  This would case the date range to be incorrectly calculated for the pre-millenium dates. The fix is to remove this transformation -- the comment says we don't have many of these, and they should probably be fixed anyway.