Fixed: Stricter matching for dates
Previously, the 'single year' match would match on a range -- so '1804-1849' would return (1804, 1804).
This was because the regex for matching this was not strict about matching the beginning and ending
of the string.

This fixes the 'simple' matches to only consider date statements that are within the beginning and the
end of the string. (More complex statements are parsed by the EDTF parser later -- this is a sort of 'pre-parser'
that tries to redirect the simplest cases away from the slower EDTF parsing path.)