Fixed: Issue with holdings and institutions
In cases where a holding had a relationship to multiple institutions, the query would actually silently
retrieve all institution records and parse the MARC, leading to an incorrect value for the city of an institution
for the holding record.

This change simplifies the holdings query and uses the siglum to look up the holding institution, but this
is further complicated by the collation used in the siglum column. A temporary fix has been added to alter
the collation of the siglum column to be accent sensitive and case sensitive so that we can use it as a key
for lookups. Hopefully this will be changed upstream in Muscat so we can remove this workaround.