Fixed: parsing dates multiple times
Previously dates were parsed multiple times, which was a problem since the date parser can be slow.

It was determined that the cause was parsing dates on the record for each incipit on that record, despite the fact that those dates would not change.

The code was fixed to use the parsed dates already on the source record.